Intern scaffolding.
A start on #9. Not sure if I like the current setup for running unit tests. I think that the jasmine test runner is much easier to work with. But I think that we should go with a full intern solution for this project as a test. Then we can port the successful parts to the boilerplate.